Pass the ParsedModule to the custom proprocessor (#3)
Currently, we pass the `ParsedSource` to the custom proprocessor.
Unfortunately, the `ParsedSource` does not contain any information
about the language extensions used in a module. In `damlc`, we would
like to issue warnings when certain language extensions are used.

This PR changes the input of the custom proprocessor to be of type
`ParsedModule`. A `ParsedModule` contains, among other things, the
`ParsedSource` as well as a `ModSummary`, which contains information
about the language extensions in use. This allows the proprocessor
to do all the things it can currenty do and to have access to the
language extensions as well.

I did deliberately not change the return type of the preprocessor
since I don't need this feature and I don't fully understand the
impact this could have if used the wrong way.

Co-authored-by: Martin Huschenbett <martin.huschenbett@posteo.me>