Pass the DynFlags to the preprocessor (#6)
Currently, we pass only the `ParsedSource` to the custom proprocessor.
Unfortunately, the `ParsedSource` does not contain any information
about the language extensions used in a module. In `damlc`, we would
like to issue warnings when certain language extensions are used.

This PR improves the situation by passing the `DynFlags` to the
preprocessor as well. The `DynFlags` contain a list of all the enabled
language extensions and hence all the information we need.