Fix the pretty printer for diagnostic locations (#7)
The error/warning location columns printed by `daml build` have been
off by one from those in DAML Studio for ages. I would argue that
DAML Studio has the ultimate authority on this since we can't change
what it displays without moving the squiggly lines.

This has consistently annoyed me when dealing with error locations in
the `damlc` integration tests but I never had the patience of figuring
out what caused the issue. Now, that I'm on a roll adding PRs to
`ghcide`, seems to be the right time to fix this.

If we merge this here, I'll make the same request with more positive
wording against upstream `ghcide`.