Stop -Werror from failing on warnings not turned on by user
We turn on certain warnings temporarily to make sure we get
diagnostics and can use them in code lenses. However, the way -Werror
works this means those warnings end up being errors which is obviously
not what we intended. This PR fixes this by removing the warnings we
turn on magically from fatal warnings if the warning is not turned on
explicitly.